requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
unstructured
faiss-cpu
streamlit
sentence-transformers
python-dotenv
beautifulsoup4
